requirements.txt
gradio==4.44.1
Pillow>=9.0.0
python-dotenv>=1.0.0 